+++ src/DependencyGraph/MatrixVertex.ts
@@ -7,7 +7,8 @@ import {AbsoluteCellRange} from '../AbsoluteCellRange'
 import {CellError, SimpleCellAddress} from '../Cell'
 import {RawCellContent} from '../CellContentParser'
 import {EmptyValue, getRawValue, InternalScalarValue} from '../interpreter/InterpreterValue'
-import {ErroredMatrix, IMatrix, Matrix, MatrixSize, NotComputedMatrix} from '../Matrix'
+import {ErroredMatrix, IMatrix, Matrix, NotComputedMatrix} from '../Matrix'
+import {MatrixSize} from '../MatrixSize'
 import {Maybe} from '../Maybe'
 import {Ast} from '../parser'
 import {ColumnsSpan, RowsSpan} from '../Span'
@@ -27,13 +28,9 @@ export class MatrixVertex {
   public static fromRange(range: AbsoluteCellRange, formula?: Ast): MatrixVertex {
     return new MatrixVertex(range.start, range.width(), range.height(), formula)
   }
-  public cellAddress: SimpleCellAddress
   public matrix: IMatrix
-  private formula: Ast | null
 
-  constructor(cellAddress: SimpleCellAddress, width: number, height: number, formula?: Ast) {
-    this.cellAddress = cellAddress
-    this.formula = formula || null
+  constructor(public cellAddress: SimpleCellAddress, width: number, height: number, private formula?: Ast) {
     this.matrix = new NotComputedMatrix(new MatrixSize(width, height))
   }
 
@@ -88,7 +85,7 @@ export class MatrixVertex {
     this.cellAddress = address
   }
 
-  public getFormula(): Ast | null {
+  public getFormula(): Maybe<Ast> {
     return this.formula
   }
 
@@ -97,11 +94,11 @@ export class MatrixVertex {
   }
 
   public isFormula(): boolean {
-    return this.formula !== null
+    return this.formula !== undefined
   }
 
   public isNumeric(): boolean {
-    return this.formula === null
+    return this.formula === undefined
   }
 
   public spansThroughSheetRows(sheet: number, startRow: number, endRow: number = startRow): boolean {
