+++ test/column-index.spec.ts
@@ -9,7 +9,7 @@ import {FunctionRegistry} from '../src/interpreter/FunctionRegistry'
 import {SimpleRangeValue} from '../src/interpreter/SimpleRangeValue'
 import {LazilyTransformingAstService} from '../src/LazilyTransformingAstService'
 import {ColumnIndex} from '../src/Lookup/ColumnIndex'
-import {Matrix, MatrixSize} from '../src/Matrix'
+import {Matrix} from '../src/Matrix'
 import {NamedExpressions} from '../src/NamedExpressions'
 import {ColumnsSpan, RowsSpan} from '../src/Span'
 import {Statistics} from '../src/statistics'
