+++ test/matrix-size-check.spec.ts
@@ -1,7 +1,7 @@
 import {CellError, ErrorType, simpleCellAddress} from '../src/Cell'
 import {Config} from '../src/Config'
 import {ErrorMessage} from '../src/error-message'
-import {checkMatrixSize, MatrixSize} from '../src/Matrix'
+import {checkMatrixSize, MatrixSize} from '../src/MatrixSize'
 import {buildEmptyParserWithCaching} from './parser/common'
 import {adr} from './testUtils'
